requirements.txt
Flask==3.1.2
flask_cors==6.0.1
python-dotenv==1.2.1
Requests==2.32.5
tabulate==0.9.0
python-dateutil==2.9.0.post0
pipreqs==0.4.13